📂 Rename:  불필요 코드 제거 및 MAIN 페이지 문구 부분 분리
